Add form to let the uesr search flights
The seed file was modified to create Flight records with a start date
from "DateTime.now" to 1 year range.

There is now a flights controller. It contains an index action. This
action gets all the flights first. It has also a variable to make a
dropdown menu for the airport codes. And another for the number of
passengers.

The root route is the index view.

The index view contains a form with some dropdowns and labels. The
dropdowns are to select the departure airport, arrival airport, number
of passengers and flight date.