Feat: Add logic to search and display flights
The Flight model contains scopes to query records based on each column
of flights table: departure airport, arrival airport, number of
passengers and flight date. It also has a method that returns all the
dates to include them in the dropdown. Finally, it also contains a
method to format a day (month/day/year).

The flights controller now correctly requires and permits the parameters
sent by the user when doing a search. In the index action, it
initializes the flights instance variable making a query that returns
zero records. It also has a new method that makes the search using the
scopes. This method only executes when there is a "commit" param.

In the index view, the form was changed to include a new dropdown that
shows only the available dates in the flights table. It uses the dates
method defined in the Flight model. This dropdown was created with the
options_from_collection_for_select method. This method receives three
arguments: a collection to iterate, a method to obtain the value from
each element and another method to obtain the value that is displayed in
each option from the dropdown.

There is also a condition in the view that when the flights instance
variable is empty, it displays a message saying that it found no flight
when searching. Otherwise, it uses a loop and in a list element it shows
the flight details (each attribute, with the date and hour formatted).