Feat: Add new Booking model
Created a new Booking model with only one field: flight_id. Done
migrations to create the bookings table and set flight_id as the foreign
key.

Set up asscociations between Booking and Flight models: a flight has
many bookings and a booking belongs to a flight.

Also created a new bookings controller that handles only the new and
create actions. The routes file is updated to allow this actions. Both
actions only create a new Booking object and assign it to an instance
variable, at the moment.

Finally, there is a view for the "new" action. It contains a p tag as a
placeholder to prevent it from being empty. This will change in another
commit.