Feat: Add form to select a flight
In the index view, there is now a new form that displays the flights
found by the other form as radio buttons. This form also contains a
hidden field that sents the num of tickets of the flight. When the user
clicks the submit button, it sends the flight_id and the num_tickets
to the "new" action of the bookings controller.

A refactor change was made with the model param in the first form. Now
it uses an instance variable called flight. This variable was defined in
the flights controller and its value is the same as it was before: it
assigns an empty Flight object.