Feat: add logic to make a booking
First, it was created and migrated a new Passenger model. It has name,
email and booking_id columns.

Set up associations. A passenger belongs to a booking. A booking has
many passengers. These two associations were made this way to allow the
building of many forms in the new sub form to make the booking. The
booking association also contains a line to allow nested parameters from
a subform.

The bookings controller contains a few new methods. It uses one to allow
the flight_id and num_tickets attributes when selecting a flight. It has
other method to obtain the num_tickets value and converts it to integer,
since every parameter is a string.

In the "new" action, it uses a loop to create empty passenger objects,
and with this, the view it creates the necessary form fields.

There is another method that allows the nested parameters and the
flight_id because the form uses another model.

In the create action, it creates another booking object using the method
above. It also creates the passengers objects thanks to the changes
mentioned above.

In the "new" view, there is a new form to make the booking. It displays
the flight data, including the flight id. Then it uses fields_for helper
method to make the subforms to enter the passengers info. There are
labels and textfields for that data and a submit button at the end.
Another important thing is that there is a hidden field to send the
flight_id value when submitting the passengers data.

Note: the create action doesn't properly redirect to show the booking
created. Also there isn't a show view for the moment.