Feat: Add show view for Booking
Changed the redirect in the "new" action to redirect to the show view
when creating a new booking. The show action searchs and returns the
booking object by the page id.

The view for the show action contains flight information and passengers
information. That's the flight departure and arrival airports, flight
date, number of passengers. And for the passengers just show the name
and email for each one.

Also allow the show action in the routes.