Feat: add confirmation emails
Installed and configured letter_opener gem to view emails in browser
instead of sending them to emails.

Created a mailer to create the emails through an action (method) that
receives a Passenger object and sets the from, to and subject parts of
and email.

Created two views to send the email: html and text. Both views have the
same content, just using different syntax for obvious reasons.

In the bookings_controller, added a loop for the passengers to call the
method that creates and make another call to send the confirmation
email for each passenger after the booking object is created.